backend-js/backend/requirements.txt
pycardano==0.9.0
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
fhir.resources==7.1.0
cryptography==41.0.5
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.1

# Atala PRISM (when SDK is available, currently placeholder)
# atala-prism-sdk==0.1.0

# Development dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
